feat : 모임 삭제 시, 다른 참가자 존재할 경우 삭제 불가 구현

+++ src/main/java/com/sosim/server/group/GroupService.java
@@ -72,6 +72,10 @@ public class GroupService {
             throw new CustomException(CodeType.NONE_ADMIN);
         }
 
+        if (participantService.getCountParticipantAtGroup(groupEntity.getId()) > 1) {
+            throw new CustomException(CodeType.NONE_ZERO_PARTICIPANT);
+        }
+
         groupEntity.delete();
     }
